Remove board class from robot
Trying to start with a simple solution first. Which is why we took out
this board class call. It was unecessary because the board class doesn't
do anything but store the pos of the robot. This pos can be easily
tracked by the robot itself. My feelings of this may change in the
future.

Other changes:
- Make the if statement more rubyish
- Ensure that if an invalid "f" is given, we say it's an invalid place.